BETA: Support for decimal values for drop list, by UnAfraid, thanks.
git-svn-id: http://svn.l2jserver.com/branches/unstable@4870 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/datatables/NpcTable.java
@@ -223,7 +223,7 @@ public class NpcTable
 					dropDat.setItemId(dropData.getInt("itemId"));
 					dropDat.setMinDrop(dropData.getInt("min"));
 					dropDat.setMaxDrop(dropData.getInt("max"));
-					dropDat.setChance(dropData.getInt("chance"));
+					dropDat.setChance(dropData.getDouble("chance"));
 					
 					int category = dropData.getInt("category");
